refactor: remove treesitter, vocab
I'll need to rethink how to approach something like this. Vocab is not
exposed by tokenizers:
- we can hardcode a count for each known model
- we could try to add it upstream in web-tokenizers

I'm also not sure about the utility - it would be nice to not decode the
entire token sequence every time to figure out if each vocab item is
relevant, but decoding is stateful and the result of one token may not
be the same as in a sequence. If we assume the only difference will be
due to merges / whitespace, we can double the vocab length and just add
a copy with whitespace added before the string.

Ultimately I think simple steering with the next prefill or static
tokens will probably be more useful. I wanted to support grammar based
sampling via tree-sitter, but tree-sitter was definitely not built for
this purpose. I think it'll be impractical even with edits to use
treesitter to parse 32k strings for each token (in Llama 2's case).

Treesitter also does things like error recovery which will make the most
common case (failing parses) take far longer. In addition, I'm not sure
we are able to tell if a source string is definitely invalid versus
could still be invalid (an incomplete input).

It's possible to still use treesitter infra and write a new parser
generator, but that'd be a massive amount of work and I'm not sure the
payoff is quite there versus just steering the first few tokens by hand,
i.e. `{ "info": ` to kick off json generation versus a JSON
grammar sampler.

And in the end, it's a tradeoff between inference errors and parser
errors - either one can be wrong.